Prepend appName to environmentName for uniquness.
While the AWS console makes it look as though Beanstalk environments are boxed inside an application this is not actually the case.

Environment names must be unique within an AWS account - so you can't have MyApp1/Production and MyApp2/Production.

As this both Wrong and Stupid I propose keeping the beanstalkify CLI as-is and magically prepending the application name to ensure uniqueness, thus: MyApp1/MyApp1-Production and MyApp2/MyApp2-Production.